Revert "Merge remote-tracking branch 'sympy/master'"
This reverts commit c738f8b679c12860e66b3787339ee855b90fb95f, reversing
changes made to 22df43e180c128587e674ff3f968e52d79175b7a.
