--- sympy/codegen/__init__.py
@@ -1,1 +1,0 @@
-from .ast import Assignment, aug_assign, CodeBlock, For
