+++ sympy/core/tests/test_args.py
@@ -133,41 +133,6 @@ def test_sympy__assumptions__sathandlers__CheckIsPrime():
     assert _test_args(CheckIsPrime(Q.positive))
     assert _test_args(CheckIsPrime(Q.positive(5)))
 
-@SKIP("abstract Class")
-def test_sympy__codegen__ast__AugmentedAssignment():
-    from sympy.codegen.ast import AugmentedAssignment
-    assert _test_args(AugmentedAssignment(x, 1))
-
-def test_sympy__codegen__ast__AddAugmentedAssignment():
-    from sympy.codegen.ast import AddAugmentedAssignment
-    assert _test_args(AddAugmentedAssignment(x, 1))
-
-def test_sympy__codegen__ast__SubAugmentedAssignment():
-    from sympy.codegen.ast import SubAugmentedAssignment
-    assert _test_args(SubAugmentedAssignment(x, 1))
-
-def test_sympy__codegen__ast__MulAugmentedAssignment():
-    from sympy.codegen.ast import MulAugmentedAssignment
-    assert _test_args(MulAugmentedAssignment(x, 1))
-
-def test_sympy__codegen__ast__DivAugmentedAssignment():
-    from sympy.codegen.ast import DivAugmentedAssignment
-    assert _test_args(DivAugmentedAssignment(x, 1))
-
-def test_sympy__codegen__ast__ModAugmentedAssignment():
-    from sympy.codegen.ast import ModAugmentedAssignment
-    assert _test_args(ModAugmentedAssignment(x, 1))
-
-def test_sympy__codegen__ast__CodeBlock():
-    from sympy.codegen.ast import CodeBlock, Assignment
-    assert _test_args(CodeBlock(Assignment(x, 1), Assignment(y, 2)))
-
-def test_sympy__codegen__ast__For():
-    from sympy.codegen.ast import For, CodeBlock, AddAugmentedAssignment
-    from sympy import Range
-    assert _test_args(For(x, Range(10), CodeBlock(AddAugmentedAssignment(y, 1))))
-
-
 @XFAIL
 def test_sympy__combinatorics__graycode__GrayCode():
     from sympy.combinatorics.graycode import GrayCode
@@ -3429,12 +3394,6 @@ def test_sympy__geometry__ellipse__Circle():
     assert _test_args(Circle((0, 1), 2))
 
 
-def test_sympy__geometry__parabola__Parabola():
-    from sympy.geometry.parabola import Parabola
-    from sympy.geometry.line import Line
-    assert _test_args(Parabola((0, 0), Line((2, 3), (4, 3))))
-
-
 @SKIP("abstract class")
 def test_sympy__geometry__line__LinearEntity():
     pass
@@ -3700,8 +3659,8 @@ def test_sympy__physics__optics__medium__Medium():
     assert _test_args(Medium('m'))
 
 
-def test_sympy__codegen__ast__Assignment():
-    from sympy.codegen.ast import Assignment
+def test_sympy__printing__codeprinter__Assignment():
+    from sympy.printing.codeprinter import Assignment
     assert _test_args(Assignment(x, y))
