+++ sympy/geometry/__init__.py
@@ -26,4 +26,3 @@ from sympy.geometry.util import are_similar, centroid, convex_hull, idiff, \
     intersection, closest_points, farthest_points
 from sympy.geometry.exceptions import GeometryError
 from sympy.geometry.curve import Curve
-from sympy.geometry.parabola import Parabola
