+++ sympy/holonomic/numerical.py
@@ -7,9 +7,9 @@ from mpmath import mp
 from sympy.holonomic.holonomic import DMFsubs
 
 
-def _evalf(func, points, derivatives=False, method='RK4'):
+def euler(func, points, derivatives=False):
     """
-    Numerical methods for numerical integration along a given set of
+    Euler method for numerical integration along a given set of
     points in the complex plane.
     """
 
@@ -17,11 +17,6 @@ def _evalf(func, points, derivatives=False, method='RK4'):
     a = ann.order
     R = ann.parent.base
     K = R.get_field()
-
-    if method == 'Euler':
-        meth = _euler
-    else:
-        meth = _rk4
 
     dmf = []
     for j in ann.listofpoly:
@@ -33,10 +28,10 @@ def _evalf(func, points, derivatives=False, method='RK4'):
     if len(y0) < a:
         raise TypeError("Not Enough Initial Conditions")
     x0 = func.x0
-    sol = [meth(red, x0, points[0], y0, a)]
+    sol = [_euler(red, x0, points[0], y0, a)]
 
     for i, j in enumerate(points[1:]):
-        sol.append(meth(red, points[i], j, sol[-1], a))
+        sol.append(_euler(red, points[i], j, sol[-1], a))
 
     if not derivatives:
         return [sympify(i[0]) for i in sol]
@@ -59,43 +54,5 @@ def _euler(red, x0, x1, y0, a):
     sol = []
     for i in range(a):
         sol.append(y_0[i] + h * f_0[i])
-
-    return sol
-
-
-def _rk4(red, x0, x1, y0, a):
-    A = sympify(x0)._to_mpmath(mp.prec)
-    B = sympify(x1)._to_mpmath(mp.prec)
-    y_0 = [sympify(i)._to_mpmath(mp.prec) for i in y0]
-    h = B - A
-
-    f_0_n = 0
-    f_1_n = 0
-    f_2_n = 0
-    f_3_n = 0
-
-    f_0 = y_0[1:]
-    for i in range(a):
-        f_0_n += sympify(DMFsubs(red[i], A))._to_mpmath(mp.prec) * y_0[i]
-    f_0.append(f_0_n)
-
-    f_1 = [y_0[i] + f_0[i]*h/2 for i in range(1, a)]
-    for i in range(a):
-        f_1_n += sympify(DMFsubs(red[i], A + h/2))._to_mpmath(mp.prec) * (y_0[i] + f_0[i]*h/2)
-    f_1.append(f_1_n)
-
-    f_2 = [y_0[i] + f_1[i]*h/2 for i in range(1, a)]
-    for i in range(a):
-        f_2_n += sympify(DMFsubs(red[i], A + h/2))._to_mpmath(mp.prec) * (y_0[i] + f_1[i]*h/2)
-    f_2.append(f_2_n)
-
-    f_3 = [y_0[i] + f_2[i]*h for i in range(1, a)]
-    for i in range(a):
-        f_3_n += sympify(DMFsubs(red[i], A + h))._to_mpmath(mp.prec) * (y_0[i] + f_2[i]*h)
-    f_3.append(f_3_n)
-
-    sol = []
-    for i in range(a):
-        sol.append(y_0[i] + h * (f_0[i]+2*f_1[i]+2*f_2[i]+f_3[i])/6)
 
     return sol
