+++ sympy/holonomic/tests/test_holonomic.py
@@ -1,7 +1,6 @@
-from sympy.holonomic import (DifferentialOperator, HolonomicFunction,
-    DifferentialOperators, from_hyper, from_meijerg, from_sympy)
+from sympy.holonomic.holonomic import DifferentialOperator, HolonomicFunction, DifferentialOperators, from_hyper
 from sympy.holonomic.recurrence import RecurrenceOperators, HolonomicSequence
-from sympy import symbols, hyper, S, sqrt, pi, exp, erf, sstr, O, I, meijerg, sin, cos
+from sympy import symbols, hyper, S, sqrt, pi, exp, erf, sstr, O, I
 from sympy import ZZ, QQ
 
 
@@ -163,23 +162,6 @@ def test_from_hyper():
     assert sstr(p.y0) == y0
     assert q.annihilator == p.annihilator
 
-def test_from_meijerg():
-    x = symbols('x')
-    R, Dx = DifferentialOperators(QQ.old_poly_ring(x), 'Dx')
-    p = from_meijerg(meijerg(([], [S(3)/2]), ([S(1)/2], [S(1)/2, 1]), x))
-    q = HolonomicFunction(2*x - 1 + (-4*x**2 + x)*Dx + 4*x**2*Dx**2 + 4*x**3*Dx**3, x, 1, \
-        [1/sqrt(pi), 1/(2*sqrt(pi)), -1/(4*sqrt(pi))])
-    assert p == q
-    p = from_meijerg(meijerg(([], []), ([0], []), x))
-    q = HolonomicFunction(1 + Dx, x, 0, 1)
-    assert p == q
-    p = from_meijerg(meijerg(([1], []), ([S(1)/2], [0]), x))
-    q = HolonomicFunction((2*x + 1)*Dx + 2*x*Dx**2, x, 1, [sqrt(pi)*erf(1), exp(-1)])
-    assert p == q
-    p = from_meijerg(meijerg(([0], [1]), ([0], []), 2*x**2))
-    q = HolonomicFunction((3*x**2 - 1)*Dx + x**3*Dx**2, x, 1, [-exp(-S(1)/2) + 1, -exp(-S(1)/2)])
-    assert p == q
-
 def test_to_Sequence():
     x = symbols('x')
     R, Dx = DifferentialOperators(ZZ.old_poly_ring(x), 'Dx')
@@ -241,70 +223,25 @@ def test_series():
     q = 1 - 3*x**2/4 - x**3/4 - 5*x**4/32 - 3*x**5/40 - 17*x**6/384 + O(x**7)
     assert p == q
 
-def test_evalf_euler():
+def test_evalf():
     x = symbols('x')
     R, Dx = DifferentialOperators(QQ.old_poly_ring(x), 'Dx')
-
-    # log(1+x)
-    p = HolonomicFunction((1 + x)*Dx**2 + Dx, x, 0, [0, 1])
-
-    # path taken is a straight line from 0 to 1, on the real axis
+    # a straight line on real axis
     r = [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1]
-    s = '0.699525841805253'  # approx. equal to log(2) i.e. 0.693147180559945
-    assert sstr(p.evalf(r, method='Euler')[-1]) == s
-
-    # path taken is a traingle 0-->1+i-->2
+    p = HolonomicFunction((1 + x)*Dx**2 + Dx, x, 0, [0, 1])
+    s = '0.699525841805253'
+    assert sstr(p.evalf(r)[-1]) == s
+    # a traingle with vertices (0, 1+i, 2)
     r = [0.1 + 0.1*I]
     for i in range(9):
         r.append(r[-1]+0.1+0.1*I)
     for i in range(10):
         r.append(r[-1]+0.1-0.1*I)
-
-    # close to the exact solution 1.09861228866811
-    # imaginary part also close to zero
     s = '1.07530466271334 - 0.0251200594793912*I'
-    assert sstr(p.evalf(r, method='Euler')[-1]) == s
-
-    # sin(x)
+    assert sstr(p.evalf(r)[-1]) == s
     p = HolonomicFunction(Dx**2 + 1, x, 0, [0, 1])
     s = '0.905546532085401 - 6.93889390390723e-18*I'
-    assert sstr(p.evalf(r, method='Euler')[-1]) == s
-
-    # computing sin(pi/2) using this method
-    # using a linear path from 0 to pi/2
-    r = [0.1]
-    for i in range(14):
-        r.append(r[-1] + 0.1)
-    r.append(pi/2)
-    s = '1.08016557252834' # close to 1.0 (exact solution)
-    assert sstr(p.evalf(r, method='Euler')[-1]) == s
-
-    # trying different path, a rectangle (0-->i-->pi/2 + i-->pi/2)
-    # computing the same value sin(pi/2) using different path
-    r = [0.1*I]
-    for i in range(9):
-        r.append(r[-1]+0.1*I)
-    for i in range(15):
-        r.append(r[-1]+0.1)
-    r.append(pi/2+I)
-    for i in range(10):
-        r.append(r[-1]-0.1*I)
-
-    # close to 1.0
-    s = '0.976882381836257 - 1.65557671738537e-16*I'
-    assert sstr(p.evalf(r, method='Euler')[-1]) == s
-
-    # cos(x)
-    p = HolonomicFunction(Dx**2 + 1, x, 0, [1, 0])
-    # compute cos(pi) along 0-->pi
-    r = [0.05]
-    for i in range(61):
-        r.append(r[-1]+0.05)
-    r.append(pi)
-    # close to -1 (exact answer)
-    s = '-1.08140824719196'
-    assert sstr(p.evalf(r, method='Euler')[-1]) == s
-
+    assert sstr(p.evalf(r)[-1]) == s
     # a rectangular path (0 -> i -> 2+i -> 2)
     r = [0.1*I]
     for i in range(9):
@@ -313,96 +250,6 @@ def test_evalf_euler():
         r.append(r[-1]+0.1)
     for i in range(10):
         r.append(r[-1]-0.1*I)
-
-    p = HolonomicFunction(Dx**2 + 1, x, 0, [1,1]).evalf(r, method='Euler')
+    p = HolonomicFunction(Dx**2 + 1, x, 0, [1,1]).evalf(r)
     s = '0.501421652861245 - 3.88578058618805e-16*I'
     assert sstr(p[-1]) == s
-
-def test_evalf_rk4():
-    x = symbols('x')
-    R, Dx = DifferentialOperators(QQ.old_poly_ring(x), 'Dx')
-
-    # log(1+x)
-    p = HolonomicFunction((1 + x)*Dx**2 + Dx, x, 0, [0, 1])
-
-    # path taken is a straight line from 0 to 1, on the real axis
-    r = [0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1]
-    s = '0.693146363174626'  # approx. equal to log(2) i.e. 0.693147180559945
-    assert sstr(p.evalf(r)[-1]) == s
-
-    # path taken is a traingle 0-->1+i-->2
-    r = [0.1 + 0.1*I]
-    for i in range(9):
-        r.append(r[-1]+0.1+0.1*I)
-    for i in range(10):
-        r.append(r[-1]+0.1-0.1*I)
-
-    # close to the exact solution 1.09861228866811
-    # imaginary part also close to zero
-    s = '1.09861574485151 + 1.36082967699958e-7*I'
-    assert sstr(p.evalf(r)[-1]) == s
-
-    # sin(x)
-    p = HolonomicFunction(Dx**2 + 1, x, 0, [0, 1])
-    s = '0.90929463522785 + 1.52655665885959e-16*I'
-    assert sstr(p.evalf(r)[-1]) == s
-
-    # computing sin(pi/2) using this method
-    # using a linear path from 0 to pi/2
-    r = [0.1]
-    for i in range(14):
-        r.append(r[-1] + 0.1)
-    r.append(pi/2)
-    s = '0.999999895088917' # close to 1.0 (exact solution)
-    assert sstr(p.evalf(r)[-1]) == s
-
-    # trying different path, a rectangle (0-->i-->pi/2 + i-->pi/2)
-    # computing the same value sin(pi/2) using different path
-    r = [0.1*I]
-    for i in range(9):
-        r.append(r[-1]+0.1*I)
-    for i in range(15):
-        r.append(r[-1]+0.1)
-    r.append(pi/2+I)
-    for i in range(10):
-        r.append(r[-1]-0.1*I)
-
-    # close to 1.0
-    s = '1.00000003415141 + 6.11940487991086e-16*I'
-    assert sstr(p.evalf(r)[-1]) == s
-
-    # cos(x)
-    p = HolonomicFunction(Dx**2 + 1, x, 0, [1, 0])
-    # compute cos(pi) along 0-->pi
-    r = [0.05]
-    for i in range(61):
-        r.append(r[-1]+0.05)
-    r.append(pi)
-    # close to -1 (exact answer)
-    s = '-0.999999993238714'
-    assert sstr(p.evalf(r)[-1]) == s
-
-    # a rectangular path (0 -> i -> 2+i -> 2)
-    r = [0.1*I]
-    for i in range(9):
-        r.append(r[-1]+0.1*I)
-    for i in range(20):
-        r.append(r[-1]+0.1)
-    for i in range(10):
-        r.append(r[-1]-0.1*I)
-
-    p = HolonomicFunction(Dx**2 + 1, x, 0, [1,1]).evalf(r)
-    s = '0.493152791638442 - 1.41553435639707e-15*I'
-    assert sstr(p[-1]) == s
-
-def test_from_sympy():
-    x = symbols('x')
-    R, Dx = DifferentialOperators(QQ.old_poly_ring(x), 'Dx')
-    p = from_sympy((sin(x)/x)**2)
-    q = HolonomicFunction(24*x + (24*x**2 + 12)*Dx + (4*x**3 + 24*x)*Dx**2 + \
-        10*x**2*Dx**3 + x**3*Dx**4, x, 1, [-cos(2)/2 + 1/2, -1 + cos(2) + sin(2), \
-        -4*sin(2) - cos(2) + 3, -12 + 14*sin(2)])
-    assert p == q
-    p = from_sympy(1/(1+x**2)**2)
-    q = HolonomicFunction(4*x + (x**2 + 1)*Dx, x, 0, 1)
-    assert p == q
