+++ sympy/printing/ccode.py
@@ -15,10 +15,8 @@ from __future__ import print_function, division
 
 from sympy.core import S
 from sympy.core.compatibility import string_types, range
-from sympy.codegen.ast import Assignment
-from sympy.printing.codeprinter import CodePrinter
+from sympy.printing.codeprinter import CodePrinter, Assignment
 from sympy.printing.precedence import precedence
-from sympy.sets.fancysets import Range
 
 # dictionary mapping sympy function to (argument_conditions, C_function).
 # Used in CCodePrinter._print_Function(self)
@@ -230,23 +228,6 @@ class CCodePrinter(CodePrinter):
             return '(*{0})'.format(name)
         else:
             return name
-
-    def _print_AugmentedAssignment(self, expr):
-        lhs_code = self._print(expr.lhs)
-        op = expr.rel_op
-        rhs_code = self._print(expr.rhs)
-        return "{0} {1} {2};".format(lhs_code, op, rhs_code)
-
-    def _print_For(self, expr):
-        target = self._print(expr.target)
-        if isinstance(expr.iterable, Range):
-            start, stop, step = expr.iterable.args
-        else:
-            raise NotImplementedError("Only iterable currently supported is Range")
-        body = self._print(expr.body)
-        return ('for ({target} = {start}; {target} < {stop}; {target} += '
-                '{step}) {{\n{body}\n}}').format(target=target, start=start,
-                stop=stop, step=step, body=body)
 
     def _print_sign(self, func):
         return '((({0}) > 0) - (({0}) < 0))'.format(self._print(func.args[0]))
