+++ sympy/printing/pretty/pretty_symbology.py
@@ -432,11 +432,6 @@ _xsym = {
     '>=':  ('>=', U('GREATER-THAN OR EQUAL TO')),
     '!=':  ('!=', U('NOT EQUAL TO')),
     ':=':  (':=', ':='),
-    '+=':  ('+=', '+='),
-    '-=':  ('-=', '-='),
-    '*=':  ('*=', '*='),
-    '/=':  ('/=', '/='),
-    '%=':  ('%=', '%='),
     '*':   ('*', U('DOT OPERATOR')),
     '-->': ('-->', U('EM DASH') + U('EM DASH') +
             U('BLACK RIGHT-POINTING TRIANGLE') if U('EM DASH')
