+++ sympy/printing/str.py
@@ -571,11 +571,6 @@ class StrPrinter(Printer):
             "==": "Eq",
             "!=": "Ne",
             ":=": "Assignment",
-            '+=': "AddAugmentedAssignment",
-            "-=": "SubAugmentedAssignment",
-            "*=": "MulAugmentedAssignment",
-            "/=": "DivAugmentedAssignment",
-            "%=": "ModAugmentedAssignment",
         }
 
         if expr.rel_op in charmap:
