+++ sympy/printing/tests/test_codeprinter.py
@@ -1,6 +1,9 @@
-from sympy.printing.codeprinter import CodePrinter
+from sympy.printing.codeprinter import CodePrinter, Assignment
 from sympy.core import symbols
 from sympy.core.symbol import Dummy
+from sympy.core.relational import Relational
+from sympy.matrices import MatrixSymbol, Matrix
+from sympy.tensor import IndexedBase, Idx
 from sympy.utilities.pytest import raises
 
 
@@ -15,6 +18,41 @@ def test_print_Dummy():
     d = Dummy('d')
     p = setup_test_printer()
     assert p._print_Dummy(d) == "d_%i" % d.dummy_index
+
+
+def test_Assignment():
+    x, y = symbols("x, y")
+    A = MatrixSymbol('A', 3, 1)
+    mat = Matrix([1, 2, 3])
+    B = IndexedBase('B')
+    n = symbols("n", integer=True)
+    i = Idx("i", n)
+    # Here we just do things to show they don't error
+    Assignment(x, y)
+    Assignment(x, 0)
+    Assignment(A, mat)
+    Assignment(A[1,0], 0)
+    Assignment(A[1,0], x)
+    Assignment(B[i], x)
+    Assignment(B[i], 0)
+    # Here we test things to show that they error
+    # Matrix to scalar
+    raises(ValueError, lambda: Assignment(B[i], A))
+    raises(ValueError, lambda: Assignment(B[i], mat))
+    raises(ValueError, lambda: Assignment(x, mat))
+    raises(ValueError, lambda: Assignment(x, A))
+    raises(ValueError, lambda: Assignment(A[1,0], mat))
+    # Scalar to matrix
+    raises(ValueError, lambda: Assignment(A, x))
+    raises(ValueError, lambda: Assignment(A, 0))
+    # Non-atomic lhs
+    raises(TypeError, lambda: Assignment(mat, A))
+    raises(TypeError, lambda: Assignment(0, x))
+    raises(TypeError, lambda: Assignment(x*x, 1))
+    raises(TypeError, lambda: Assignment(A + A, mat))
+    raises(TypeError, lambda: Assignment(B, 0))
+
+    assert Relational(x, y, ':=') == Assignment(x, y)
 
 def test_print_Symbol():
