+++ sympy/printing/tests/test_fcode.py
@@ -1,8 +1,8 @@
-from sympy import (sin, cos, atan2, log, exp, gamma, conjugate, sqrt,
-    factorial, Integral, Piecewise, Add, diff, symbols, S, Float, Dummy, Eq,
-    Range, Catalan, EulerGamma, E, GoldenRatio, I, pi, Function, Rational, Integer, Lambda)
+from sympy import sin, cos, atan2, log, exp, gamma, conjugate, sqrt, \
+    factorial, Integral, Piecewise, Add, diff, symbols, S, Float, Dummy, Eq
+from sympy import Catalan, EulerGamma, E, GoldenRatio, I, pi
+from sympy import Function, Rational, Integer, Lambda
 
-from sympy.codegen import For, Assignment
 from sympy.core.relational import Relational
 from sympy.logic.boolalg import And, Or, Not, Equivalent, Xor
 from sympy.printing.fcode import fcode, FCodePrinter
@@ -665,13 +665,3 @@ def test_Matrix_printing():
         "      M(1, 3) = cos(q(3, 1))\n"
         "      M(2, 3) = 5\n"
         "      M(3, 3) = 0")
-
-
-def test_fcode_For():
-    x, y = symbols('x y')
-
-    f = For(x, Range(0, 10, 2), [Assignment(y, x * y)])
-    sol = fcode(f)
-    assert sol == ("      do x = 0, 10, 2\n"
-                   "         y = x*y\n"
-                   "      end do")
